Book Store Dto id type Long to İnt

+++ src/main/java/com/bookStoreApp/repo/BookRepository.java
@@ -3,5 +3,5 @@ package com.bookStoreApp.repo;
 import com.bookStoreApp.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<Book, Long> {
+public interface BookRepository extends JpaRepository<Book, Integer> {
 }
