+++ src/main/java/com/bookStoreApp/repo/BookStoreRepository.java
@@ -3,5 +3,5 @@ package com.bookStoreApp.repo;
 import com.bookStoreApp.entity.BookStore;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookStoreRepository extends JpaRepository<BookStore, Long> {
+public interface BookStoreRepository extends JpaRepository<BookStore, Integer> {
 }
